Working(?) OpenMP
Has file reading, uses the matrix as a double array. Currently assumes that the file it reads is set in a name (matrix) by us; prints: every node's shortest path (or 9999 if it doesn't exist), the shortest path in the matrix (excluding source), and the destination's length (if exists). Does not display the path to the destination or ask for the source/destination.